bug: fix circular dependency
Problem: The function createCredentialsFileWatcher and its dependencies
were written in the filesystemUtilities.ts and causing a circular
dependency.
Solution: Move the function to credentialsFile.ts. Update the imports.
